+++ packages/core/src/plugin.ts
@@ -1,6 +1,8 @@
-import * as STG from '.'
+import * as stg from '.'
 
-export type STG = typeof STG
+export type STG = typeof stg
+
+export type Wrapped<T, S extends any[] = []> = T | ((...args: S) => T)
 
 export type PluginFunction<T = any> = (stg: STG, options?: T) => any
 
@@ -8,12 +10,19 @@ export type Plugin<T = any> = PluginFunction<T> | {
   install: PluginFunction<T>
 }
 
-export function use<T>(plugin: Plugin<T>, options?: T) {
+const installedPlugin = new Set<PluginFunction>()
+
+export function use<T>(plugin: Plugin<T>, options?: T, once = true) {
+  let install: PluginFunction
   if (typeof plugin === 'function') {
-    return plugin(STG, options)
-  } else if (typeof plugin.install === 'function') {
-    return plugin.install(STG, options)
+    install = plugin
+  } else if (plugin && typeof plugin.install === 'function') {
+    install = plugin.install
   } else {
-    throw new Error(`Not a valid plugin.`)
+    throw new TypeError(`Not a valid plugin. A plugin should be a function or an object with an "install" method.`)
   }
+  if (installedPlugin.has(install)) return
+  const result = install(stg, options)
+  if (once) installedPlugin.add(install)
+  return result
 }