+++ packages/touhou/src/index.ts
@@ -1,23 +1,11 @@
-import { STG } from '@stg/core'
-
-import flaming from './templates/flaming'
-import large from './templates/large'
-import medium from './templates/medium'
-import small from './templates/small'
-import ellipse from './templates/ellipse'
-
-export function install(stg: STG) {
-  stg.defineTemplate({
-    flaming,
-    large,
-    medium,
-    small,
-    ellipse,
-  })
-}
-
+import templates from './templates'
+import bullets from './bullets'
+import th12 from './touhou/th12'
 import th15 from './touhou/th15'
 
 export {
+  templates,
+  bullets,
+  th12,
   th15,
 }