+++ packages/utils/src/config.ts
@@ -1,0 +1,9 @@
+export interface UtilsConfig {
+  cacheSize: number
+  minValue: number
+}
+
+export default {
+  cacheSize: 4096,
+  minValue: 1e-4,
+} as UtilsConfig