+++ packages/utils/src/index.ts
@@ -1,5 +1,6 @@
 import * as random from './random'
 import memorize from './memorize'
+import _config from './config'
 import color from './color'
 import math from './math'
 
@@ -8,4 +9,7 @@ export {
   math,
   memorize,
   color,
+  _config,
 }
+
+export * from './config'