+++ packages/web-stg/src/library.js
@@ -1,8 +1,6 @@
 import * as stg from '@stg/core'
 import * as touhou from '@stg/touhou'
 
-stg.use(touhou)
-
 export * from '@stg/core'
 export * from '@stg/utils'
 export { touhou }